requirements.txt
ttkbootstrap
Pillow
playsound==1.2.2
matplotlib
platformdirs
requests
